users_micro/requirements.txt
fastapi
uvicorn[standard]
# For Pydantic
pydantic[email]
# For authentication
python-multipart
python-jose[cryptography]
passlib[bcrypt]
# For database
sqlalchemy
psycopg2-binary
# Environment file
python-dotenv
# Additional dependencies
requests
pycryptodome
google-auth
# For UUID support
uuid
# For file uploads
python-multipart
# For image processing (optional)
Pillow

python-magic

aiofiles
dotenv
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0

opencv-python>=4.8.0 
Pillow>=10.0.0


numpy>=1.24.0  